requirements.txt
asgiref==3.7.2
boto3==1.28.25
botocore==1.31.25
certifi==2023.7.22
Django==4.2.4
django-minio-backend==3.5.0
django-minio-storage==0.5.6
django-storages==1.13.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.7
inflection==0.5.1
jmespath==1.0.1
minio==7.1.15
packaging==23.1
psycopg2==2.9.7
PyJWT==2.8.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0.1
s3transfer==0.6.1
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.16
